arcface_onnx_service/requirements.txt
fastapi
uvicorn[standard]
insightface
numpy
pillow
psycopg2-binary
onnxruntime
python-multipart